update to set the default name of runsettings-file for .net-tests to "runsettings.xml"
background: you want to run the testcases using the runtestcases.py-script. this works like before if you name the runsettingsfile accordingly. but you may also run the testcases in visual studio. this does not work with usual .runsettings-files like in grylibrary because tasksforcommonprojectstructure uses dotnet-coverage. dotnet-coverage only uses a part of the usual "big" runsettings-file. using this part only in a .runsettings-file is not compatible with visual studio in the sence that you cannot run or debug testcases in visual studio anymore. so for debugging-purposes the runsettings-file must have another extension (now: xml) when running the new dotnet-coverage-command.